Seperated auth handlers and task handlers
